Chore: Rename component files to PascalCase Updated component filenames to use PascalCase
