+++ src/components/login-form/index.tsx
@@ -5,14 +5,14 @@ import { checkLoginStatus, storeLoginToken } from '../../utils/auth';
 import { useNavigate } from 'react-router-dom';
 import { UserBasicLoginData } from '../../types/user';
 import { validateEmail, validatePassword } from '../../utils/validators';
-import { Heading } from '../common/heading';
-import { Wrapper } from '../common/wrapper';
-import { Label } from '../common/label';
-import { StatusBlock } from '../common/statusBlock';
-import { Cta } from '../common/cta';
-import { Spinner } from '../common/spinner';
-import { Form } from '../common/form';
-import { Input } from '../common/input';
+import { Heading } from '../common/Heading';
+import { Wrapper } from '../common/Wrapper';
+import { Label } from '../common/Label';
+import { StatusBlock } from '../common/StatusBlock';
+import { Cta } from '../common/Cta';
+import { Spinner } from '../common/Spinner';
+import { Form } from '../common/Form';
+import { Input } from '../common/Input';
 
 export default function LoginForm() {
   const navigate = useNavigate()
