+++ src/pages/users/index.tsx
@@ -3,11 +3,11 @@ import { UserManagement } from "./management"
 import UsersList from "./list"
 import { SpecificUserPage } from "./user"
 import { ThemeProvider } from "styled-components"
-import { Hero } from "../../components/common/hero"
-import { Heading } from "../../components/common/heading"
-import { Wrapper } from "../../components/common/wrapper"
-import { Menu } from "../../components/common/menu"
-import { Link } from "../../components/common/link"
+import { Hero } from "../../components/common/Hero"
+import { Heading } from "../../components/common/Heading"
+import { Wrapper } from "../../components/common/Wrapper"
+import { Menu } from "../../components/common/Menu"
+import { Link } from "../../components/common/Link"
 import theme from "../../themes/default"
